+++ tests/model_evaluation/test_visualizations.py
@@ -8,7 +8,7 @@ from pathlib import Path
 import numpy as np
 import pandas as pd
 import pytest
-from sklearn.metrics import f1_score, roc_auc_score
+from sklearn.metrics import f1_score, recall_score, roc_auc_score
 
 from psycop_model_training.model_eval.base_artifacts.plots.base_charts import (
     plot_basic_chart,
@@ -19,11 +19,15 @@ from psycop_model_training.model_eval.base_artifacts.plots.feature_importance im
 from psycop_model_training.model_eval.base_artifacts.plots.performance_by_age import (
     plot_performance_by_age,
 )
+from psycop_model_training.model_eval.base_artifacts.plots.performance_by_sex import (
+    plot_performance_by_sex,
+)
 from psycop_model_training.model_eval.base_artifacts.plots.performance_over_time import (
     plot_auc_by_time_from_first_visit,
     plot_metric_by_calendar_time,
     plot_metric_by_cyclic_time,
     plot_metric_by_time_until_diagnosis,
+    plot_recall_by_calendar_time,
 )
 from psycop_model_training.model_eval.base_artifacts.plots.precision_recall import (
     plot_precision_recall,
@@ -103,7 +107,14 @@ def test_plot_bar_chart(synth_eval_dataset: EvalDataset):
 def test_plot_performance_by_age(synth_eval_dataset: EvalDataset):
     plot_performance_by_age(
         eval_dataset=synth_eval_dataset,
-        save_path=PROJECT_ROOT / "test.png",
+        save_path=PROJECT_ROOT / "test_performance_plot_by_age.png",
+    )
+
+
+def test_plot_performance_by_sex(synth_eval_dataset: EvalDataset):
+    plot_performance_by_sex(
+        eval_dataset=synth_eval_dataset,
+        save_path=PROJECT_ROOT / "test_performance_plot_by_sex.png",
     )
 
 
@@ -119,6 +130,19 @@ def test_plot_performance_by_calendar_time(
         eval_dataset=synth_eval_dataset,
         bin_period=bin_period,
         metric_fn=roc_auc_score,
+        save_path=PROJECT_ROOT / "test.png",
+    )
+
+
+def test_plot_recall_by_calendar_time(
+    synth_eval_dataset: EvalDataset,
+):
+    plot_recall_by_calendar_time(
+        eval_dataset=synth_eval_dataset,
+        pred_proba_percentile=[0.8, 0.9, 0.95],
+        bins=list(range(0, 1460, 180)),
+        y_limits=(0, 0.5),
+        save_path=PROJECT_ROOT / "test_recall_by_calendar_time.png",
     )